requirements.txt
fastapi
uvicorn
sqlalchemy>=2
passlib            # 用 PBKDF2-SHA256，不依賴 bcrypt
pyjwt
pydantic
email-validator
python-multipart
gunicorn
